Overhaul how we use the wicket-merged-resources library.
Before we maintained a list of commons resources in the application and then used that list both for initializing merged-resources and for attaching each resource to the header in `BasePage`. This was a lot of code.

Simplify this by introducing a new helper class, `MergedResourceBuilder`. This takes care of setting all sorts of reasonable defaults and also automatically attaches the merged resources to a page of your choice, like `BasePage`, for example.

This means that a whole bunch of code from the application and base page can be removed. All that is needed to set up merged resources is some concise builder code in the application init.

Need to add javadoc.

Addresses #14.

git-svn-id: http://opensource.55minutes.com/svn/java/trunk/wicket@107 3a26569e-5468-0410-bb2f-b495330926e5