Another retooling of the merged resources implementation.
I took out the instantiation listener, because I think it would be problematic in practice. Wicket does not guarantee the order that these listeners are executed, so if you used the 55 listener in conjunction with the merge-resources annotations, the results could be unpredictable.

Instead, make the `build()` method also responsible for returning a header contributor that will contribute all the merged CSS/JS resources. This contributor can then be explicitly passed to the pages that need it. In our archetype and example code, we use it in the `BasePage` so that it is common across all pages. Although this is not as elegant as the listener magic, it is still DRY, and more robust.

Also add lots of javadoc. Still needs some test coverage. Addresses #14.

git-svn-id: http://opensource.55minutes.com/svn/java/trunk/wicket@108 3a26569e-5468-0410-bb2f-b495330926e5