Make login flow more explicit
Remove the option to automatically stay on the same page after the user
clicks the login link and successfully authenticates. There is no easy way
to accomplish this in Wicket without the URL changing, which is not what
we want.

Instead, by default redirect to the home page upon successful login.
Developers can customize this by specifying a destination page when
constructing the login link.