Improve feedback during important security events
A good practice is to redirect the user to a known accessible page
(e.g. home page) with a useful feedback message rather than showing a
scary "access denied" error page.

Make this the default in wicket-shiro, while also supporting the more
traditional scary 403 error page approach.

Apply the same reasonable-default-with-customization pattern for
feedback messages for the "login required" and "logged off" situations.
In our archetype, provide a sample `WicketApplication.properties` so
that developers can easily see how these feedback messages can be
customized.

Also tweak the archetype code a bit so that every page has a feedback
panel, thus allowing the feedback messages to work out of the box.

Closes #10